update treatment of cloning and build tags.
